Refactor getMissingDependencies() to accept plugins.
CoreUtils.getMissingDependencies() now takes a list of
plugins and the plugin registry. It uses this to
first determine the list of dependencies, and second
to determine which dependencies are not satisfied.

TODO: The newly supplied plugins can satisfy their
own dependencies, and this is not taken into account.

WIP: Comment out tests and update signature of getMissingDependencies()

WIP: Uncomment & fix one getMissingDependencies() test

WIP: Update plugin in getMissingDependencies() to use interface.

Fix remaining commented getMissingDependencies() test.