SFX-99: Add Core module with registration mechanism (#5)
Core is the backbone of the entire SF-X system. It manages plugins and
allows them to communicate with each other. The core package defines
interfaces for plugins.

Exported from the `@sfx/core` package are the `Core` class and the
`Plugin`, `PluginMetadata` and `PluginRegistry` interfaces.

A registration mechanism is currently implemented. Unregistration will
be added separately. The interface and semantics of the process are
documented in the class and interface.