SFX-99: Implement blanket unregistration in Core (#7)
Core now implements an `unregisterAll` function that will unregister all
plugins. Unregistering one or more plugins without unregistering all of
them is not yet supported.

Also added is an optional `unregister` hook to the plugin interface that
is called when the plugin is being unregistered. Other plugins are still
available to this plugin during this hook; they are only removed after
all plugins to be unregistered have had their `unregister` hook called.