SFX-198: Implement plugin unregistration (#27)
A new method Core.unregister has been added to unregister a subset of
plugins. It takes in a list of plugins to unregister and unregisters
them. A new dependency graph structure has been added to keep track of
dependency relationships, and Core.register has been modified to
populate that graph.

The graph is represented as an adjacency list, which is represented as
an object where the keys are plugin names and the values are arrays of
dependers (i.e. plugins that list the first plugin as a dependency). A
number of dependency graph-related helper functions have also been
created.