SFX-248: Update repo to pull in events and interfaces from @sfx/events package (#19)
Normalize all events and event payloads to use and conform to the
exports of the @sfx/events package. All instances of CustomEvent are
also properly typed.

The "module" property in the package-specific package.json files have
been removed so that the package can be properly consumed in a project
that uses Webpack as a bundler.

In addition, the TypeScript file regex has been corrected. The original
regex (genreated by webpack-cli) matched any file that had at least one
character in its name. The correct regex should match any file that ends
with `.ts` or `.tsx` and nothing else.