SFX-385: Allow for passing custom product transformer (#30)
This required a new `productTransformer` property on both SaytDriver and SearchDriver. A function can be passed into the constructor to set it. The transformer converts any `Record` objects returned from a search into a format desired for rendering. The View Layer can then use this simplified/transformed data to render (say) a grid of products.

Additionally, the drivers were updated to accept a generic type P which represents the interface to which a product should conform.

Lastly, the README was updated to include usage examples.