SFX-158: Add Cache and Cache Driver plugins (#34)
The cache is meant for plugins to store data. In the current GB Elements
use case, this cache serves as an event cache. The intention is for the
drivers to store the latest API response (more accurately, the payload
of response events) in the cache so that components can request them on
mount. The scope may grow to storing persistent data, like search
configuration. Concretely, the cache plugin exposes a Map that plugins
can interact with.

The cache driver serves as an event-based interface to the cache.