ELE-247: Add quick start package (#45)
A new package, `@groupby/elements-quickstart`, creates an instance of `Core` and registers all the plugins on it. The plugins are configurable through an options object accepted by the `quickStart` function.

A new global is exposed by Webpack, `GbElementsLogic`, that contains the `quickStart` function and all the other Logic layer entities.